lksctp-tools: package libwithsctp with the shell script using it
withsctp is a simple shell script that just LD_PRELOADs libwithsctp,
and there's little use for libwithsctp unless combined with that
script. So put the library in the -withsctp package.

packageqa now complains, but I cannot figure out why it didn't
complain about the same thing when the library was in the main
package.