skaware: add (big-endian) powerpc support
sysdeps mostly identical the those for arm. Compile-tested only so far,
but there's no reason to believe this shouldn't just work. Most of the
sysdeps depend on (features provided by) kernel and libc, and not on the
underlying architecture. The exception being the endianness, which
annoyingly is not strictly determined by being a 32-bit ppc. However, we
don't have support for any little-endian ppc in
oe-lite (cf. lib/oelite/arch.py in meta/core), so this should be ok.